--- intelligence/ts/src/engines/caching/storage.ts
@@ -1,27 +1,0 @@
-// Copyright 2025 Flower Labs GmbH. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// =============================================================================
-
-export interface CachedEntry {
-  engineModel: string;
-  timestamp: number;
-}
-
-export interface CachedMapping {
-  mapping: Record<string, CachedEntry>;
-}
-export abstract class CacheStorage {
-  abstract getItem(key: string): Promise<CachedEntry | null>;
-  abstract setItem(key: string, value: string | null): Promise<void>;
-}
